update gmsh to version 4.1.2

+++ scripts/gmsh_default_build.sh
@@ -14,9 +14,19 @@ $CMAKE .. -DCMAKE_INSTALL_PREFIX=${TwoPiRoot} \
           -DCMAKE_INSTALL_RPATH=${TwoPiRoot}  \
           -DENABLE_BUILD_DYNAMIC=1            \
 	  -DENABLE_OS_SPECIFIC_INSTALL=0      \
+          -DENABLE_WRAP_PYTHON=1              \
+          -DENABLE_BUILD_SHARED=1             \
           -DCMAKE_INSTALL_NAME_DIR=${TwoPiRoot}/lib 
 
 $MAKE $MAKEOPT
 $MAKE install
 
-ln -s $TwoPiRoot/lib/gmsh.py gmsh.py
+# add link to gmsh.py. it seems like there is two possible locations
+cd $TwoPiRoot/lib/python2.7/site-packages
+if [ -f $TwoPiRoot/lib/gmsh.py ]; then
+   ln -s $TwoPiRoot/lib/gmsh.py gmsh.py
+fi
+if [ -f $TwoPiRoot/lib64/gmsh.py ]; then
+   ln -s $TwoPiRoot/lib64/gmsh.py gmsh.py
+fi
+
